usePadarias hook (refatoracao)
