+++ frontend/src/routes/PadariasList/index.tsx
@@ -1,30 +1,10 @@
 import PadariaCard from "../../components/PadariaCard";
 import "./styles.scss"
-import axiosInstance from '../../axios';
-import { useEffect, useState } from "react";
-
-interface Padaria {
-    id: number,
-    nome_fantasia: string
-}
+import { usePadarias } from './usePadarias'
 
 const PadariasList = () => {
 
-    const [padarias, setPadarias] = useState<Padaria[]> ([]);
-
-    useEffect(() => {
-    (
-        async () => {
-        axiosInstance.get('/')
-        .then((response) => {
-            setPadarias(response.data.padarias);
-        })
-        .catch(() => {
-            alert("Ih Serjão, sujou!")
-        })
-        }
-    )();
-    },[])
+   const { padarias } = usePadarias();
 
 
     return <div id="padarias-list">
